requirements.txt
# requirements.txt
torch==2.0.0
torchvision==0.15.0
timm==0.9.2
albumentations==1.3.1
pandas==2.0.3
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.3.0
tqdm==4.65.0